// cms/src/store/ducks/messages/types.ts
export enum MessagesTypes {
  LOAD_REQUEST_FETCH_INBOX = '@messages/fetch_inbox/LOAD_REQUEST',
  LOAD_SUCCESS_FETCH_INBOX = '@messages/fetch_inbox/LOAD_SUCCESS',
  LOAD_FAILURE_FETCH_INBOX = '@messages/fetch_inbox/LOAD_FAILURE',
  LOAD_REQUEST_FETCH_SEND = '@messages/fetch_send/LOAD_REQUEST',
  LOAD_SUCCESS_FETCH_SEND = '@messages/fetch_send/LOAD_SUCCESS',
  LOAD_FAILURE_FETCH_SEND = '@messages/fetch_send/LOAD_FAILURE',
  LOAD_REQUEST_SUBMIT = '@messages/submit/LOAD_REQUEST',
  LOAD_SUCCESS_SUBMIT = '@messages/submit/LOAD_SUCCESS',
  LOAD_FAILURE_SUBMIT = '@messages/submit/LOAD_FAILURE',
  LOAD_REQUEST_SUBMITATTACHMENT = '@messages/submitAttachment/LOAD_REQUEST',
  LOAD_SUCCESS_SUBMITATTACHMENT = '@messages/submitAttachment/LOAD_SUCCESS',
  LOAD_FAILURE_SUBMITATTACHMENT = '@messages/submitAttachment/LOAD_FAILURE',
  LOAD_REQUEST_GET_ATTACHMENT = '@messages/getAttachment/LOAD_REQUEST',
  LOAD_SUCCESS_GET_ATTACHMENT = '@messages/getAttachment/LOAD_SUCCESS',
  LOAD_FAILURE_GET_ATTACHMENT = '@messages/getAttachment/LOAD_FAILURE',
  LOAD_REQUEST_FETCH_ID = '@messages/fetchId/LOAD_REQUEST',
  LOAD_SUCCESS_FETCH_ID = '@messages/fetchId/LOAD_SUCCESS',
  LOAD_FAILURE_FETCH_ID = '@messages/fetchId/LOAD_FAILURE',
  LOAD_REQUEST_DELETE = '@messages/delete/LOAD_REQUEST',
  LOAD_SUCCESS_DELETE = '@messages/delete/LOAD_SUCCESS',
  LOAD_FAILURE_DELETE = '@messages/delete/LOAD_FAILURE',
  LOAD_REQUEST_MESSAGES_TYPES = '@messages/MESSAGES_TYPES/LOAD_REQUEST',
  LOAD_SUCCESS_MESSAGES_TYPES = '@messages/MESSAGES_TYPES/LOAD_SUCCESS',
  LOAD_FAILURE_MESSAGES_TYPES = '@messages/MESSAGES_TYPES/LOAD_FAILURE',
  LOAD_REQUEST_MESSAGES_STATUS = '@messages/MESSAGES_STATUS/LOAD_REQUEST',
  LOAD_SUCCESS_MESSAGES_STATUS = '@messages/MESSAGES_STATUS/LOAD_SUCCESS',
  LOAD_FAILURE_MESSAGES_STATUS = '@messages/MESSAGES_STATUS/LOAD_FAILURE',
  LOAD_REQUEST_LIST_USERS_TO_SEND = '@messages/LIST_USERS_TO_SEND/LOAD_REQUEST',
  LOAD_SUCCESS_LIST_USERS_TO_SEND = '@messages/LIST_USERS_TO_SEND/LOAD_SUCCESS',
  LOAD_FAILURE_LIST_USERS_TO_SEND = '@messages/LIST_USERS_TO_SEND/LOAD_FAILURE',
  LOAD_REQUEST_LIST_USERS_FILTERED = '@messages/LIST_USERS_FILTERED/LOAD_REQUEST',
  LOAD_SUCCESS_LIST_USERS_FILTERED = '@messages/LIST_USERS_FILTERED/LOAD_SUCCESS',
  LOAD_FAILURE_LIST_USERS_FILTERED = '@messages/LIST_USERS_FILTERED/LOAD_FAILURE',

  LOAD_REQUEST_LIST_INSTITUTIONS = '@messages/LIST_INSTITUTIONS/LOAD_REQUEST',
  LOAD_SUCCESS_LIST_INSTITUTIONS = '@messages/LIST_INSTITUTIONS/LOAD_SUCCESS',
  LOAD_FAILURE_LIST_INSTITUTIONS = '@messages/LIST_INSTITUTIONS/LOAD_FAILURE',
  LOAD_REQUEST_LIST_LOCAL_UNITS = '@messages/LIST_LOCAL_UNITS/LOAD_REQUEST',
  LOAD_SUCCESS_LIST_LOCAL_UNITS = '@messages/LIST_LOCAL_UNITS/LOAD_SUCCESS',
  LOAD_FAILURE_LIST_LOCAL_UNITS = '@messages/LIST_LOCAL_UNITS/LOAD_FAILURE',
  LOAD_REQUEST_LIST_PROFILES = '@messages/LIST_PROFILES/LOAD_REQUEST',
  LOAD_SUCCESS_LIST_PROFILES = '@messages/LIST_PROFILES/LOAD_SUCCESS',
  LOAD_FAILURE_LIST_PROFILES = '@messages/LIST_PROFILES/LOAD_FAILURE',
  LOAD_REQUEST_LIST_SUPERINTENDENCES = '@messages/LIST_SUPERINTENDENCES/LOAD_REQUEST',
  LOAD_SUCCESS_LIST_SUPERINTENDENCES = '@messages/LIST_SUPERINTENDENCES/LOAD_SUCCESS',
  LOAD_FAILURE_LIST_SUPERINTENDENCES = '@messages/LIST_SUPERINTENDENCES/LOAD_FAILURE',

  SORT_FILTER = '@messages/sort/SORT_FILTER',
}